fully_generated/high-fidelity-generative-compression-master/high-fidelity-generative-compression-master/requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==1.0.0
annotated-types==0.7.0
astunparse==1.6.3
autograd==1.7.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dlib==19.24.6
easydict==1.13
efficientnet-pytorch==0.7.1
eval_type_backport==0.2.2
filelock==3.17.0
flatbuffers==25.1.21
fonttools==4.55.5
fsspec==2024.12.0
gast==0.4.0
gdown==5.2.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
imutils==0.5.4
insightface==0.7.3
Jinja2==3.1.5
joblib==1.4.2
keras==2.11.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
munch==4.0.0
namex==0.0.8
networkx==3.2.1
numpy==1.20.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pretrainedmodels==0.7.4
prettytable==3.14.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
retinaface-pytorch==0.0.8
rich==13.9.4
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.6
simsimd==6.2.1
six==1.17.0
soupsieve==2.6
stringzilla==3.11.3
sympy==1.13.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.10.0
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
